In investigating Issue #114, I determined that when the legacy parser encountered an "else" statement, it would eat the next statement. This caused this code to fail to see the class definition:
if x
else
  class A
    def silly; end
  end
end

The class A would be unnoticed, but the def silly would be seen and be created at the
top level.

By fixing this, I expose 4 failing tests that were passing because of the bug. Luckily,
they're features that exist in the 1.9 handlers but not yet in the 1.8.7 handlers, so
once those features are ported, then the 1.8 code gets a boost as well as correctly-passing
specs!